Renamed dto folder to remotemodel
