+++ app/src/main/java/acn/assessment/weatherapp/viewmodels/MainViewModel.kt
@@ -3,8 +3,8 @@ package acn.assessment.weatherapp.viewmodels
 import android.annotation.SuppressLint
 import androidx.lifecycle.ViewModel
 import acn.assessment.weatherapp.service.data.repositories.WeatherRepository
-import acn.assessment.weatherapp.service.data.dto.CurrentWeather
-import acn.assessment.weatherapp.service.data.dto.FullWeather
+import acn.assessment.weatherapp.service.data.remotemodel.CurrentWeather
+import acn.assessment.weatherapp.service.data.remotemodel.FullWeather
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
